feat(openapi): add comments and error convert file
