+++ db/entities/jsonschema.go
@@ -1,7 +1,6 @@
 package entities
 
 import (
-	_ "embed"
 	"encoding/json"
 	"fmt"
 	"github.com/getkin/kin-openapi/openapi3"
@@ -12,7 +11,7 @@ import (
 	"reflect"
 )
 
-var Schemas openapi3.Schemas
+var schemas openapi3.Schemas
 
 func LoadOpenAPI(bytes []byte) {
 	loader := openapi3.NewLoader()
@@ -27,7 +26,7 @@ func LoadOpenAPI(bytes []byte) {
 	); err != nil {
 		panic(fmt.Errorf("OpenAPI document validation failed: %w", err))
 	}
-	Schemas = doc.Components.Schemas
+	schemas = doc.Components.Schemas
 }
 
 func getStructName(i interface{}) string {
@@ -41,10 +40,11 @@ func getStructName(i interface{}) string {
 	return ""
 }
 
-func NewEntity[T any](defaultSet func(*T)) (T, map[string]interface{}) {
+// NewEntity new a default entity by OpenAPI schema definition
+func NewEntity[T any](defaultSet func(*T)) T {
 	var entity T
-	jsonObject := make(map[string]interface{})
-	val, err := Schemas.JSONLookup(getStructName(entity))
+	defaultObject := make(map[string]interface{})
+	val, err := schemas.JSONLookup(getStructName(entity))
 	if err != nil {
 		panic(fmt.Errorf("schema not found for entity %s: %w", getStructName(entity), err))
 	}
@@ -53,7 +53,7 @@ func NewEntity[T any](defaultSet func(*T)) (T, map[string]interface{}) {
 		panic(fmt.Errorf("type is not schema for entity %s", getStructName(entity)))
 	}
 
-	_ = schema.VisitJSON(jsonObject,
+	_ = schema.VisitJSON(defaultObject,
 		openapi3.MultiErrors(),
 		openapi3.VisitAsRequest(),
 		openapi3.DisableReadOnlyValidation(),
@@ -64,13 +64,16 @@ func NewEntity[T any](defaultSet func(*T)) (T, map[string]interface{}) {
 		}),
 	)
 
-	b, _ := json.Marshal(jsonObject)
-	_ = json.Unmarshal(b, &entity)
-	return entity, jsonObject
+	b, _ := json.Marshal(defaultObject)
+	if err = json.Unmarshal(b, &entity); err != nil {
+		panic(fmt.Errorf("failed to new default entity %w", err))
+	}
+	return entity
 }
 
+// Validate validate the entity by OpenAPI schema defintion
 func Validate[T any](entity T) error {
-	val, err := Schemas.JSONLookup(getStructName(entity))
+	val, err := schemas.JSONLookup(getStructName(entity))
 	if err != nil {
 		return fmt.Errorf("schema not found for entity %s: %w", getStructName(entity), err)
 	}
@@ -100,40 +103,10 @@ func Validate[T any](entity T) error {
 	return nil
 }
 
-func ValidateByMap[T any](entity *T, jsonObject map[string]interface{}) error {
-	val, err := Schemas.JSONLookup(getStructName(entity))
-	if err != nil {
-		return fmt.Errorf("schema not found for entity %s: %w", getStructName(entity), err)
-	}
-	schema, ok := val.(*openapi3.Schema)
-	if !ok {
-		return fmt.Errorf("type is not schema for entity %s", getStructName(entity))
-	}
-
-	err = schema.VisitJSON(jsonObject,
-		openapi3.MultiErrors(),
-		openapi3.VisitAsRequest(),
-		openapi3.DisableReadOnlyValidation(),
-	)
-	switch err := err.(type) {
-	case nil:
-	case openapi3.MultiError:
-		issues := openapi.ConvertError(err, "@body")
-		jsonIssues := utils.ConvertJSONPaths(issues)
-		return errs.NewValidateFieldsError(errs.ErrRequestValidate, jsonIssues)
-	default:
-		return err
-	}
-	b, _ := json.Marshal(jsonObject)
-	if err := json.Unmarshal(b, entity); err != nil {
-		return errs.NewValidateError(err)
-	}
-	return nil
-}
-
+// UnmarshalAndValidate unmarshal the request body as entity, combine the default value and validate it
 func UnmarshalAndValidate[T any](r io.ReadCloser, entity *T, defaultSet func(*T)) error {
-	jsonObject := make(map[string]interface{})
-	val, err := Schemas.JSONLookup(getStructName(entity))
+	objectMap := make(map[string]interface{})
+	val, err := schemas.JSONLookup(getStructName(entity))
 	if err != nil {
 		panic(fmt.Errorf("schema not found for entity %s: %w", getStructName(entity), err))
 	}
@@ -142,7 +115,7 @@ func UnmarshalAndValidate[T any](r io.ReadCloser, entity *T, defaultSet func(*T)
 		panic(fmt.Errorf("type is not schema for entity %s", getStructName(entity)))
 	}
 
-	_ = schema.VisitJSON(jsonObject,
+	_ = schema.VisitJSON(objectMap,
 		openapi3.MultiErrors(),
 		openapi3.VisitAsRequest(),
 		openapi3.DisableReadOnlyValidation(),
@@ -153,14 +126,14 @@ func UnmarshalAndValidate[T any](r io.ReadCloser, entity *T, defaultSet func(*T)
 		}),
 	)
 
-	b, _ := json.Marshal(jsonObject)
+	b, _ := json.Marshal(objectMap)
 	_ = json.Unmarshal(b, entity)
 
-	if err := json.NewDecoder(r).Decode(&jsonObject); err != nil {
+	if err := json.NewDecoder(r).Decode(&objectMap); err != nil {
 		return err
 	}
 
-	err = schema.VisitJSON(jsonObject,
+	err = schema.VisitJSON(objectMap,
 		openapi3.MultiErrors(),
 		openapi3.VisitAsRequest(),
 		openapi3.DisableReadOnlyValidation(),
@@ -174,7 +147,7 @@ func UnmarshalAndValidate[T any](r io.ReadCloser, entity *T, defaultSet func(*T)
 	default:
 		return err
 	}
-	b, _ = json.Marshal(jsonObject)
+	b, _ = json.Marshal(objectMap)
 	if err := json.Unmarshal(b, entity); err != nil {
 		return errs.NewValidateError(err)
 	}
